src/requirements.txt
Mako==0.7.0
Markdown==2.1.1
MarkupSafe==0.15
PyRSS2Gen==1.0.0
argparse==1.2.1
boto==2.5.1
bottle==0.10.9
certifi==0.0.8
chardet==1.0.1
dropbox==1.4
gevent==0.13.7
greenlet==0.3.4
oauth==1.0.1
oauthlib==0.1.3
py-bcrypt==0.2
pyasn1==0.1.3
requests==0.13.1
rsa==3.1.1
simplejson==2.5.2
wsgiref==0.1.2
